+++ events.mjs
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 // Backbone.Events
 // ---------------
