+++ model.mjs
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { Events } from './events.mjs';
-import { extend, wrapError, urlError, addUnderscoreMethods } from './helpers.mjs';
+import { extend, wrapError, urlError, addMethodsToObject } from './helpers.mjs';
 
 // Backbone.Model
 // --------------
@@ -367,22 +367,4 @@ _.extend(Model.prototype, Events, {
 var modelMethods = {keys: 1, values: 1, pairs: 1, invert: 1, pick: 0,
     omit: 0, chain: 1, isEmpty: 1};
 
-// Mix in each Underscore method as a proxy to `Collection#models`.
-
-_.each([
-    [Model, modelMethods, 'attributes']
-], function(config) {
-    var Base = config[0],
-        methods = config[1],
-        attribute = config[2];
-
-    Base.mixin = function(obj) {
-        var mappings = _.reduce(_.functions(obj), function(memo, name) {
-            memo[name] = 0;
-            return memo;
-        }, {});
-        addUnderscoreMethods(Base, obj, mappings, attribute);
-    };
-
-    addUnderscoreMethods(Base, _, methods, attribute);
-});
+addMethodsToObject([Model, modelMethods, 'attributes']);
